atv-dsm4/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
pymysql
cryptography
pydantic
flasgger
python-dotenv